Extract enemy movement
